--- cypress/e2e/tests/migration/ci/ci.test.ts
@@ -1,30 +1,0 @@
-/*
-Copyright © 2021 the Konveyor Contributors (https://konveyor.io/)
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-/// <reference types="cypress" />
-
-import {
-    archetypeCRUD,
-    assessReviewAndAnalyzeApplication,
-    businessServiceCRUD,
-    jobFunctionCRUD,
-} from "./ci-utils";
-
-describe(["@ci"], "Sanity UI tests", () => {
-    businessServiceCRUD();
-    jobFunctionCRUD();
-    archetypeCRUD();
-    assessReviewAndAnalyzeApplication();
-});
